+++ Sources/KeyboardKit/Keyboard/KeyboardContext.swift
@@ -146,13 +146,13 @@ public class KeyboardContext: ObservableObject {
      The text document proxy that is currently active.
      */
     @Published
-    public var textDocumentProxy: UITextDocumentProxy = PreviewTextDocumentProxy()
-    
+    public var originalTextDocumentProxy: UITextDocumentProxy = PreviewTextDocumentProxy()
+
     /**
      The text document proxy that is currently active.
      */
     @Published
-    public var originalTextDocumentProxy: UITextDocumentProxy = PreviewTextDocumentProxy()
+    public var textDocumentProxy: UITextDocumentProxy = PreviewTextDocumentProxy()
     
     /**
      The text input mode of the input controller.
@@ -317,11 +317,12 @@ public extension KeyboardContext {
         if screenSize != controller.screenSize {
             screenSize = controller.screenSize
         }
-        if textDocumentProxy === controller.textDocumentProxy {} else {
-            textDocumentProxy = controller.textDocumentProxy
-        }
+
         if originalTextDocumentProxy === controller.originalTextDocumentProxy {} else {
             originalTextDocumentProxy = controller.originalTextDocumentProxy
+        }
+        if textDocumentProxy === controller.textDocumentProxy {} else {
+            textDocumentProxy = controller.textDocumentProxy
         }
         if textInputMode != controller.textInputMode {
             textInputMode = controller.textInputMode